Allow constant interpreter to skip builtin instructions from profiling instrumentation
