+++ lib/SILOptimizer/Transforms/AccessEnforcementReleaseSinking.cpp
@@ -139,6 +139,7 @@ static bool isBarrier(SILInstruction *inst) {
     case BuiltinValueKind::TypePtrAuthDiscriminator:
     case BuiltinValueKind::GlobalStringTablePointer:
     case BuiltinValueKind::COWBufferForReading:
+    case BuiltinValueKind::IntInstrprofIncrement:
       return false;
 
     // Handle some rare builtins that may be sensitive to object lifetime
